refactor(account): move API helpers to pages/account/api
