+++ src/pages/account/api/signInWithGoogle.ts
@@ -3,7 +3,7 @@ import { GoogleAuthProvider, linkWithPopup, signInWithCredential, type User, typ
 
 import { auth } from "@/shared/firebase";
 
-export const loginGoogle = async (): Promise<User> => {
+export const signInWithGoogle = async (): Promise<User> => {
   const { currentUser } = auth;
   if (!currentUser?.isAnonymous) throw new Error("Anonymous user is required before Google sign-in");
