+++ src/pages/account/api/signOutCurrentUser.spec.ts
