+++ src/pages/account/api/signOutCurrentUser.ts
