+++ src/pages/account/model/actions/signIn.ts
@@ -1,9 +1,9 @@
 import type { AccountPageStore } from "../types";
-import { loginGoogle } from "./loginGoogle";
+import { signInWithGoogle } from "../../api/signInWithGoogle";
 import { runAccountAction } from "./runAccountAction";
 
 export function signIn(store: AccountPageStore, isMounted: () => boolean): Promise<void> {
-  return runAccountAction(loginGoogle, store, isMounted, {
+  return runAccountAction(signInWithGoogle, store, isMounted, {
     operation: "signIn",
     success: "Signed in.",
     failure: "Unable to sign in.",
